reworking uvs to be floats

+++ src/rendering/shapes.hpp
@@ -11,15 +11,15 @@
 
 typedef struct {
   float x, y, z;
-  int16_t u, v;
+  float u, v;
 } vertex;
 
 #define QUAD_VERTICES \
 { \
-  { 0.0f, 0.0f, 0.5f, 1, 1 }, \
-  { 1.0f, 0.0f, 0.5f, 0, 1 }, \
-  { 1.0f, 1.0f, 0.5f, 0, 0 }, \
-  { 0.0f, 1.0f, 0.5f, 1, 0 } \
+  { 0.0f, 0.0f, 0.5f, 1.0f, 1.0f }, \
+  { 1.0f, 0.0f, 0.5f, 0.0f, 1.0f }, \
+  { 1.0f, 1.0f, 0.5f, 0.0f, 0.0f }, \
+  { 0.0f, 1.0f, 0.5f, 1.0f, 0.0f } \
 }
 
 #define QUAD_INDICES { 0, 1, 2, 0, 2, 3 }
